Split validity into multiple files
