+++ rundmcmc/validity/validity.py
@@ -1,49 +1,3 @@
-"""
-Collection of constraint functions for the validation step in RunDMCMC.
-
-=========================== ==============================================
-Helper classes
-==========================================================================
-Validator                   Collection of constraints
-Bounds                      Bounds on numeric constraints
-UpperBounds                 Upper bounds on numeric constraints
-LowerBounds                 Lower bounds on numeric constraints
-SelfConfiguringUpperBound   Automatic upper bounds on numeric constraints
-SelfConfiguringLowerBound   Automatic lower bounds on numeric constraints
-WithinPercentRangeOfBounds  Percentage bounds for numeric constraints
-==========================================================================
-
-|
-
-============================================== ==============================================
-Boolean constraint functions
-=============================================================================================
-no_worse_L1_reciprocal_polsby_popper            Lower bounded L1-reciprocal Polsby-Popper
-no_worse_L_minus_1_reciprocal_polsby_popper     Lower bounded L(-1)-reciprocal Polsby-Popper
-single_flip_contiguous                          Contiguity of districts after single flips
-contiguous                                      Contiguity of districts with NetworkX methods
-no_more_disconnected                            No more disconnected districts than initially
-no_vanishing_districts                          No districts may be completely consumed
-=============================================================================================
-
-Each new step proposed to the chain is passed off to the "validator" functions
-here to determine whether or not the step is valid. If it is invalid (breaks
-contiguity, for instance), then the step is immediately rejected.
-
-The signature of a validator function should be as follows::
-
-    def validator(partition):
-        # check if valid
-        # ...
-        return is_valid
-
-That is, validators take in a :class:`~rundmcmc.partition.Partition` instance,
-and should return whether or not the instance is valid according to their
-rules. Many top-level functions following this signature in this module are
-examples of this.
-
-"""
-
 import collections
 import logging
 import random
@@ -53,6 +7,8 @@ import networkx.algorithms.shortest_paths.weighted as nx_path
 from networkx import NetworkXNoPath
 
 from rundmcmc.updaters import CountySplit
+from rundmcmc.validity.bounds import (SelfConfiguringLowerBound, SelfConfiguringUpperBound,
+                                      Bounds)
 
 logger = logging.getLogger(__name__)
 
@@ -83,155 +39,6 @@ class Validator:
 
         # all constraints are satisfied
         return True
-
-
-class Bounds:
-    """
-    Wrapper for numeric-validators to enforce upper and lower limits.
-
-    This class is meant to be called as a function after instantiation; its
-    return is ``True`` if the numeric validator is within set limits, and
-    ``False`` otherwise.
-
-    """
-    def __init__(self, func, bounds):
-        """
-        :func: Numeric validator function. Should return an iterable of values.
-        :bounds: Tuple of (lower, upper) numeric bounds.
-        """
-        self.func = func
-        self.bounds = bounds
-
-    def __call__(self, *args, **kwargs):
-        lower, upper = self.bounds
-        values = self.func(*args, **kwargs)
-        return lower <= min(values) and max(values) <= upper
-
-    @property
-    def __name__(self):
-        return "Bounds({})".format(self.func.__name__)
-
-
-class UpperBound:
-    """
-    Wrapper for numeric-validators to enforce upper limits.
-
-    This class is meant to be called as a function after instantiation; its
-    return is ``True`` if the numeric validator is within a set upper limit,
-    and ``False`` otherwise.
-
-    """
-    def __init__(self, func, bound):
-        """
-        :func: Numeric validator function. Should return a comparable value.
-        :bounds: Comparable upper bound.
-        """
-        self.func = func
-        self.bound = bound
-
-    def __call__(self, *args, **kwargs):
-        return self.func(*args, **kwargs) <= self.bound
-
-    @property
-    def __name__(self):
-        return "UpperBound({})".format(self.func.__name__)
-
-
-class LowerBound:
-    """
-    Wrapper for numeric-validators to enforce lower limits.
-
-    This class is meant to be called as a function after instantiation; its
-    return is ``True`` if the numeric validator is within a set lower limit,
-    and ``False`` otherwise.
-
-    """
-    def __init__(self, func, bound):
-        """
-        :func: Numeric validator function. Should return a comparable value.
-        :bounds: Comparable lower bound.
-        """
-        self.func = func
-        self.bound = bound
-
-    def __call__(self, *args, **kwargs):
-        return self.func(*args, **kwargs) >= self.bound
-
-    @property
-    def __name__(self):
-        return "LowerBound({})".format(self.func.__name__)
-
-
-class SelfConfiguringUpperBound:
-    """
-    Wrapper for numeric-validators to enforce automatic upper limits.
-
-    When instantiated, the initial upper bound is set as the initial value of
-    the numeric-validator.
-
-    This class is meant to be called as a function after instantiation; its
-    return is ``True`` if the numeric validator is within a set upper limit,
-    and ``False`` otherwise.
-
-    """
-    def __init__(self, func):
-        """
-        :func: Numeric validator function.
-        """
-        self.func = func
-        self.bound = None
-
-    def __call__(self, partition):
-        if not self.bound:
-            self.bound = self.func(partition)
-            return self.__call__(partition)
-        else:
-            return self.func(partition) <= self.bound
-
-
-class SelfConfiguringLowerBound:
-    """
-    Wrapper for numeric-validators to enforce automatic lower limits.
-
-    When instantiated, the initial lower bound is set as the initial value of
-    the numeric-validator minus some configurable ε.
-
-    This class is meant to be called as a function after instantiation; its
-    return is ``True`` if the numeric validator is within a set lower limit,
-    and ``False`` otherwise.
-
-    """
-    def __init__(self, func, epsilon=0.05):
-        """
-        :func: Numeric validator function.
-        :epsilon: Initial "wiggle room" that the validator allows.
-        """
-        self.func = func
-        self.bound = None
-        self.epsilon = epsilon
-
-    def __call__(self, partition):
-        if not self.bound:
-            self.bound = self.func(partition) - self.epsilon
-            return self.__call__(partition)
-        else:
-            return self.func(partition) >= self.bound
-
-
-class WithinPercentRangeOfBounds:
-    def __init__(self, func, percent):
-        self.func = func
-        self.percent = float(percent) / 100.
-        self.lbound = None
-        self.ubound = None
-
-    def __call__(self, partition):
-        if not (self.lbound and self.ubound):
-            self.lbound = self.func(partition) * (1.0 - self.percent)
-            self.ubound = self.func(partition) * (1.0 + self.percent)
-            return True
-        else:
-            return self.lbound <= self.func(partition) <= self.ubound
 
 
 def L1_reciprocal_polsby_popper(partition):
